--- Assets/Scripts/PlayerInteraction.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 4450a1b26aaa7d743a600a1533b69626
